fix of #519
