+++ src/packages/EmbedIO/Net/Internal/SystemHttpResponse.cs
@@ -1,6 +1,8 @@
 ﻿using System;
+using System.Collections;
 using System.IO;
 using System.Net;
+using System.Runtime.Serialization;
 using System.Text;
 using EmbedIO.Utilities;
